Scans of my derivation of the new formulation
This derivation assumes a totally generic function of the data (misfit,
vTv, etc) and generic function of the parameters (Marquardt, Tikhonov,
etc) subject to the fully non-linear f(d, p) = 0 using Lagrange
multipliers.

The derivation is like we did in
https://github.com/pinga-lab/inverse-problems : expand in Taylor series
of 2nd order, then gradient of expansion equal to null (i.e., Newton's
method). Then, I disregard all 2nd derivatives of f(d, p) like in the
Gauss approximation of the Hessian.

This results in a delta p that simplifies to the regularized parametric
cases and the traditional implicit formulation from Geodesy.
Furthermore, we can add any regularization we want (to parameters and
data) and use true norm of residuals. All we need to do is calculate the
Hessian and the gradient of each new constraint we want. Perfect for
implementing in Fatiando.