Convergence is better using proper weights
For the data, use the norm of the data squared. For the parameters,
using the survey area and data norm worked well. Still some things to
work out regarding convergence of the algorithm, how well things fit the
Euler equation, etc. But for now, we can get a decent fit and check the
misfit.