Finish implementation and make iterator methods
This way we can step through the inversion 1 iteration at a time. Not
useful later on but for the paper this will be great to show
convergence, etc. Notebook runs a synthetic and shows progress through
iterations. Perhaps using the merit function isn't really needed since
it's still dropping long after the position stabilizes.