Apply Euler inversion to real data from Rio (#17)
I had another go at applying the inversion to the real data from Rio with the nice intrusions and dikes. I used @indiauppal's code but modified to use the data from the archive in the repository (removing the bad tie line data and expanding the region). I tried the application on gridded data and on the raw line data (derivatives calculated with equivalent sources). The gridded data would either result in very deep sources (for smooth solutions) or dikes interpreted as many dipoles (for shallow solutions). In the end, I think the line data worked best so that's what I kept. I ran the windowed version on the full dataset and also isolated 4 anomalies and ran the non-windowed inversion on them as well. The results are good enough. The windowed version gets most of the dike on the bottom right but the others it either misses or thinks they're SI 2 or 3. That's reasonable since those always have a lot of interference. For the larger anomalies it mostly gets them right with SI 3 (but also a mix of other things). The spread is expected of a windowed algorithm. The individual anomaly inversions show that the SI estimate works on the real data. The lesson I guess is that the windowing procedure sucks we should probably isolate anomalies by hand. But that's for another paper.