Add an implementation of finite-difference Euler deconv (#25)
The FD Euler method uses differences between the data and the data point
at the center of a window to eliminate the constant base level. With
this, they can estimate SI as well. Implement the method and test of the
first synthetic to see how it compares. The result is worse than Euler
deconv when there is noise and an interfering source.