Lock the environment and update reproduction instructions (#36)
Use conda-lock to create a cross-platform file that has the exact
versions of everything used to generate our results. This file can be
used to create an environment identical to ours. Update the
REPRODUCING.md instructions to use this instead of the `environment.yml`
file. Rerun all results with the new locked environment to make sure
everything is up to date. Modify the Makefile to make sure dependencies
are correct and add a rule to generate the lock file.