Create a diff with the submitted version (#42)
We need to submit a manuscript with the marked changes. Add a command to
the Makefile that generates this PDF using `latexdiff`. Unfortunately,
it can't be installed through conda on Windows (even though the
[feedstock](https://github.com/conda-forge/latexdiff-feedstock) says
it's `noarch`). This is because of a complicated Perl build issue on
Windows (see https://github.com/conda-forge/perl-feedstock/issues/58).
Add instructions for people trying to reproduce this.